Add a bit of fake world gen

+++ Expanse/src/com/expanse/celestialobject/BeltObject.java
@@ -4,9 +4,9 @@ import com.expanse.resources.Resources;
 
 public class BeltObject extends CelestialObject{
 	
-	protected float orbitThickness;
+	protected double orbitThickness;
 	
-	public BeltObject(String name, Resources remainingR, float orbitalR, float orbitalT){
+	public BeltObject(String name, Resources remainingR, double orbitalR, double orbitalT){
 		this.owningPlayer = -1;
 		this.objectName = name;
 		this.remainingResources = remainingR;
