+++ Expanse/src/com/expanse/celestialobject/TerrestrialPlanetaryMassBody.java
@@ -4,12 +4,12 @@ import com.expanse.resources.Resources;
 
 public class TerrestrialPlanetaryMassBody extends PlanetaryMassBody{
 	protected Resources coreResources;
-	protected float breathableProportion; //0-1
-	protected float atmosphericDensity; //Earth atmospheres
-	protected float surfaceTemperature; //Kelvin, optimally 25oC, 298oK?
+	protected double breathableProportion; //0-1
+	protected double atmosphericDensity; //Earth atmospheres
+	protected double surfaceTemperature; //Kelvin, optimally 25oC, 298oK?
 	//TO simplify things absolute zero is -273oC, ignoring the .15
 	
-	public TerrestrialPlanetaryMassBody(String name, Resources remainingR, Resources coreR, float orbitalR, float r, float m, float breathP, float atmoDensity, float surfaceTemp){
+	public TerrestrialPlanetaryMassBody(String name, Resources remainingR, Resources coreR, double orbitalR, double r, double m, double breathP, double atmoDensity, double surfaceTemp){
 		this.owningPlayer = -1;
 		this.objectName = name;
 		this.remainingResources = remainingR;
